Add methods to ClientManager class

+++ headers/BusinessPartner.h
@@ -4,16 +4,16 @@
 class BusinessPartner {
     public:
         BusinessPartner();
-        BusinessPartner(int id, std::string description, int legal_id, Adress adress, int phone, std::string email);
+        BusinessPartner(int id, std::string description, long long int legal_id, Adress adress, int phone, std::string email);
         void setId(int id);
         void setDescription(std::string description);
-        void setLegalId(int legal_id);
+        void setLegalId(long long int legal_id);
         void setAdress(Adress adress);
         void setPhone(int phone);
         void setEmail(std::string email);
         int getId();
         std::string getDescription();
-        int getLegalId();
+        long long int getLegalId();
         Adress getAdress();
         int getPhone();
         std::string getEmail();
